refactor(presets): rename es6 to es2015
