+++ es6.js
@@ -1,8 +1,3 @@
 'use strict';
 
-const commonRules = require('./lib/common-rules');
-const es6Rules = require('./lib/es6-rules');
-
-module.exports = {
-    rules: Object.assign({}, commonRules, es6Rules)
-};
+module.exports = require('./es2015');
