+++ examples/bad/es2015.js
