+++ index.js
@@ -1,5 +1,3 @@
 'use strict';
 
-const es6 = require('./es6');
-
-module.exports = es6;
+module.exports = require('./es2015');
