+++ lib/es2015-rules.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = {
-    /* ECMAScript 6 ======================================================================== */
+    /* ECMAScript 2015 ===================================================================== */
     /* http://eslint.org/docs/rules/#ecmascript-6
     /* ===================================================================================== */
     'constructor-super': 'error',
